+++ src/Controller/PageController.php
@@ -6,11 +6,14 @@ use App\Entity\Page;
 use App\Repository\PageRepository;
 use App\Repository\ChapterRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use App\Middleware\NovelRelationMiddleware;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Bundle\SecurityBundle\Security as SecurityAuth;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class PageController extends AbstractController
@@ -19,19 +22,27 @@ class PageController extends AbstractController
     private $pageRepo;
     private $chapterRepo;
 
-    public function __construct(EntityManagerInterface $em, PageRepository $pageRepo, ChapterRepository $chapterRepo){
+    public function __construct(EntityManagerInterface $em, PageRepository $pageRepo, ChapterRepository $chapterRepo, NovelRelationMiddleware $novelRelationMiddleware , private SecurityAuth $security){
         $this->em = $em;
         $this->pageRepo = $pageRepo;
         $this->chapterRepo = $chapterRepo;
+        $this->novelRelationMiddleware = $novelRelationMiddleware;
     }
 
-    #[Route('/page', methods: ['POST'])]
+    #[Route('/page', methods: ['POST']), Security("is_granted('IS_AUTHENTICATED_FULLY')")]
     public function createPage(Request $request, SerializerInterface $serializer){
         $data = json_decode($request->getContent(),true);
         $page = new Page();
+        $chapter = $this->chapterRepo->find($data["chapter"]);
+        $novel = $chapter->getNovel();
+        $user = $this->security->getUser();
+
+        if (!$this->novelRelationMiddleware->isUserAuthorized($novel, $user)) {
+            return $this->json(['error' => 'Vous éte pas l\'author de cette novel du coup vous ne pouver pas le supprimer : '. $novel->getId()], 404);
+        }
         $page->setContent($data["content"]);
         $page->setHtml($data["html"]);
-        $page->setChapter($this->chapterRepo->find($data["chapter"]));
+        $page->setChapter($chapter);
         $this->em->persist($page);
         $this->em->flush();
         $json = $serializer->serialize($page, 'json', ['groups' => 'page:read']);
@@ -48,13 +59,21 @@ class PageController extends AbstractController
         return new JsonResponse($json, 200, [], true);
     }
 
-    #[Route('/page/{id}', methods: ['PUT'])]
+    #[Route('/page/{id}', methods: ['PUT']), Security("is_granted('IS_AUTHENTICATED_FULLY')")]
     public function updatePage(int $id, Request $request, SerializerInterface $serializer){
         $data = json_decode($request->getContent(),true);
         $page = $this->pageRepo->find($id);
         if (!$page) {
             return $this->json(['error' => 'No found id: '. $id], 404);
         }
+
+        $novel = $page->getChapter()->getNovel();
+        $user = $this->security->getUser();
+
+        if (!$this->novelRelationMiddleware->isUserAuthorized($novel, $user)) {
+            return $this->json(['error' => 'Vous éte pas l\'author de cette novel du coup vous ne pouver pas le supprimer : '. $novel->getId()], 404);
+        }
+
         $page->setContent($data["content"]);
         $page->setHtml($data["html"]);
         $chapter = $this->chapterRepo->find($data["chapter"]);
@@ -65,12 +84,20 @@ class PageController extends AbstractController
         return new JsonResponse($json, 200, [], true);
     }
 
-    #[Route('/page/{id}', methods: ['DELETE'])]
+    #[Route('/page/{id}', methods: ['DELETE']), Security("is_granted('IS_AUTHENTICATED_FULLY')")]
     public function deletePage(int $id){
         $page = $this->pageRepo->find($id);
         if (!$page) {
             return $this->json(['error' => 'No found id: '. $id], 404);
         }
+
+        $novel = $page->getChapter()->getNovel();
+        $user = $this->security->getUser();
+
+        if (!$this->novelRelationMiddleware->isUserAuthorized($novel, $user)) {
+            return $this->json(['error' => 'Vous éte pas l\'author de cette novel du coup vous ne pouver pas le supprimer : '. $novel->getId()], 404);
+        }
+
         $this->em->remove($page);
         $this->em->flush();
         return new Response("no content", 204);
